requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
minio>=7.1.0
pandas>=1.5.0
sentence-transformers>=2.2.2
chromadb>=0.3.21
fastapi>=0.95.0
uvicorn>=0.22.0
pytest>=7.3.0
apache-airflow==2.9.3
pyarrow>=10.0.0
typing_extensions>=4.10.0
more-itertools>=9.0.0
faiss-cpu>=1.7.2